NF: moves RenderBrowserQA where it's useful
