+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionTask.java
@@ -1097,74 +1097,6 @@ public class CollectionTask<Progress, Result> extends BaseAsyncTask<Void, Progre
         }
     }
 
-
-    public static class RenderBrowserQA implements TaskDelegate<Integer, Pair<CardBrowser.CardCollection<CardBrowser.CardCache>, List<Long>>> {
-        private final CardBrowser.CardCollection<CardBrowser.CardCache> mCards;
-        private final Integer mStartPos;
-        private final Integer mN;
-        private final int mColumn1Index;
-        private final int mColumn2Index;
-
-
-        public RenderBrowserQA(CardBrowser.CardCollection<CardBrowser.CardCache> cards, Integer mStartPos, Integer n, int column1Index, int column2Index) {
-            this.mCards = cards;
-            this.mStartPos = mStartPos;
-            this.mN = n;
-            this.mColumn1Index = column1Index;
-            this.mColumn2Index = column2Index;
-        }
-
-
-        public Pair<CardBrowser.CardCollection<CardBrowser.CardCache>, List<Long>> task(@NonNull Collection col, @NonNull ProgressSenderAndCancelListener<Integer> collectionTask) {
-            Timber.d("doInBackgroundRenderBrowserQA");
-
-            List<Long> invalidCardIds = new ArrayList<>();
-            // for each specified card in the browser list
-            for (int i = mStartPos; i < mStartPos + mN; i++) {
-                // Stop if cancelled
-                if (collectionTask.isCancelled()) {
-                    Timber.d("doInBackgroundRenderBrowserQA was aborted");
-                    return null;
-                }
-                if (i < 0 || i >= mCards.size()) {
-                    continue;
-                }
-                CardBrowser.CardCache card;
-                try {
-                    card = mCards.get(i);
-                }
-                catch (IndexOutOfBoundsException e) {
-                    //even though we test against card.size() above, there's still a race condition
-                    //We might be able to optimise this to return here. Logically if we're past the end of the collection,
-                    //we won't reach any more cards.
-                    continue;
-                }
-                if (card.isLoaded()) {
-                    //We've already rendered the answer, we don't need to do it again.
-                    continue;
-                }
-                // Extract card item
-                try {
-                    // Ensure that card still exists.
-                    card.getCard();
-                } catch (WrongId e) {
-                    //#5891 - card can be inconsistent between the deck browser screen and the collection.
-                    //Realistically, we can skip any exception as it's a rendering task which should not kill the
-                    //process
-                    long cardId = card.getId();
-                    Timber.e(e, "Could not process card '%d' - skipping and removing from sight", cardId);
-                    invalidCardIds.add(cardId);
-                    continue;
-                }
-                // Update item
-                card.load(false, mColumn1Index, mColumn2Index);
-                float progress = (float) i / mN * 100;
-                collectionTask.doProgress((int) progress);
-            }
-            return new Pair<>(mCards, invalidCardIds);
-        }
-    }
-
     public static class RebuildCram implements TaskDelegate<Void, StudyOptionsFragment.DeckStudyData> {
         public StudyOptionsFragment.DeckStudyData task(@NonNull Collection col, @NonNull ProgressSenderAndCancelListener<Void> collectionTask) {
             Timber.d("doInBackgroundRebuildCram");
